requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==44.0.3
psycopg-binary==3.2.10
distlib==0.3.9
distro==1.9.0
Django==5.1.4
django-environ==0.12.0
djangorestframework==3.16.1
drf-yasg==1.21.11
et_xmlfile==2.0.0
filelock==3.17.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.11.0
numpy==2.3.2
numpy-financial==1.0.0
openai==2.0.1
openpyxl==3.1.5
oracledb==3.1.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.6
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.3
requests==2.32.4
rsa==4.9.1
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.29.1